implemented session in data layer

+++ app/src/main/java/com/lsrv/copypastedetector/data/repositories/Repository.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.snapshots.SnapshotStateList
 
 interface Repository<T> {
     suspend fun refresh()
-    suspend fun get(id: Int): T
+    suspend fun get(id: Long): T
     suspend fun getAll(): SnapshotStateList<T>
     suspend fun insert(item: T): Long
     suspend fun insertAll(vararg items: T)
